+++ CHANGELOG.md
@@ -7,9 +7,11 @@
 
 ### Changed
 - Custom fields’ search keywords are now indexed for entry drafts.
+- Updated Yii to 2.0.43.
 
 ### Fixed
 - Fixed a bug where custom fields’ search keywords weren’t getting indexed for newly-created entries. ([#9699](https://github.com/craftcms/cms/issues/9699))
+- Fixed a bug where user email addresses would validate if they had IDNA ASCII characters in the local part, even though Swiftmailer doesn’t allow it. ([#9096](https://github.com/craftcms/cms/issues/9096))
 
 ## 3.7.8 - 2021-08-06